+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.3.0 - 13-07-2019
+
+### Added
+- ***AngleStick Class*** - AngleStick is a angle join which will prevent points from bending at certain degrees, helpfull for making ragdolls
+- Entity.js - **addAngleStick**
+- Utils.js - **radians, degrees**
+- Vector.js - **angle, angle2, rotateBy**
+- Verly.js - **setDPI()** function which will improve the rendering in mobile devices
+- **Added Basic Unit Testing**
+
 ## v1.2.1 - 11-07-2019
 
 ### Added